test: move --jsx-* dev-runtime coverage to its own file
The existing jsx-production.test.ts describe.concurrent block spawns a
heavy fixture (react-dom render + Bun.build + dynamic import) that races
the 5s default timeout under debug+ASAN. Keeping the new --jsx-* CLI flag
tests in a separate file gives them a clean pass/fail independent of that
pre-existing timing.
